fixed stacks combined pipeline mode
