+++ servicecatalog_factory/template_builder/pipeline/__init__.py
@@ -1,0 +1,2 @@
+#  Copyright 2022 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+#  SPDX-License-Identifier: Apache-2.0
